pyproject.toml
[project]
name = "canslim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.30.1",
    "finnhub-python>=2.4.20",
    "flask>=3.1.2",
    "flask-apscheduler>=1.13.1",
    "flask-cors>=6.0.2",
    "google-adk>=1.25.0",
    "litellm>=1.81.10",
    "openai>=2.20.0",
    "pandas-market-calendars>=5.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "pytz>=2024.1",
    "requests>=2.32.5",
    "supabase>=2.28.0",
    "werkzeug>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-env>=1.1.3",
    "pytest-mock>=3.12.0",
]
